Move result types to impl package
