+++ subprojects/execution/src/main/java/org/gradle/internal/execution/history/BeforeExecutionState.java
@@ -17,7 +17,7 @@
 package org.gradle.internal.execution.history;
 
 import com.google.common.collect.ImmutableSortedMap;
-import org.gradle.internal.execution.SnapshotResult;
+import org.gradle.internal.execution.steps.SnapshotResult;
 import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint;
 import org.gradle.internal.snapshot.FileSystemSnapshot;
