+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/BroadcastChangingOutputsStep.java
@@ -19,7 +19,6 @@ package org.gradle.internal.execution.steps;
 import com.google.common.collect.ImmutableList;
 import org.gradle.api.file.FileCollection;
 import org.gradle.internal.execution.OutputChangeListener;
-import org.gradle.internal.execution.Result;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.file.TreeType;
