+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/BuildCacheStep.java
@@ -26,7 +26,6 @@ import org.gradle.caching.internal.controller.BuildCacheCommandFactory.LoadMetad
 import org.gradle.caching.internal.controller.BuildCacheController;
 import org.gradle.caching.internal.origin.OriginMetadata;
 import org.gradle.internal.Try;
-import org.gradle.internal.execution.CurrentSnapshotResult;
 import org.gradle.internal.execution.ExecutionOutcome;
 import org.gradle.internal.execution.ExecutionResult;
 import org.gradle.internal.execution.OutputChangeListener;
