+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/CachingResult.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution;
+package org.gradle.internal.execution.steps;
 
+import org.gradle.internal.execution.ExecutionEngine;
 import org.gradle.internal.execution.caching.CachingState;
 
 public interface CachingResult extends UpToDateResult, ExecutionEngine.Result {
