+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/DeferredExecutionAwareStep.java
@@ -19,7 +19,6 @@ package org.gradle.internal.execution.steps;
 import org.gradle.cache.Cache;
 import org.gradle.internal.Try;
 import org.gradle.internal.execution.DeferredExecutionHandler;
-import org.gradle.internal.execution.Result;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.UnitOfWork.Identity;
