+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/IdentityCacheStep.java
@@ -21,7 +21,6 @@ import org.gradle.internal.Cast;
 import org.gradle.internal.Try;
 import org.gradle.internal.execution.DeferredExecutionHandler;
 import org.gradle.internal.execution.ExecutionResult;
-import org.gradle.internal.execution.Result;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.UnitOfWork.Identity;
