+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/SkipUpToDateStep.java
@@ -21,11 +21,9 @@ import com.google.common.collect.ImmutableSortedMap;
 import org.apache.commons.lang.StringUtils;
 import org.gradle.caching.internal.origin.OriginMetadata;
 import org.gradle.internal.Try;
-import org.gradle.internal.execution.CurrentSnapshotResult;
 import org.gradle.internal.execution.ExecutionOutcome;
 import org.gradle.internal.execution.ExecutionResult;
 import org.gradle.internal.execution.UnitOfWork;
-import org.gradle.internal.execution.UpToDateResult;
 import org.gradle.internal.execution.history.AfterPreviousExecutionState;
 import org.gradle.internal.snapshot.FileSystemSnapshot;
 import org.slf4j.Logger;
