+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/ValidateStep.java
@@ -20,7 +20,6 @@ import com.google.common.collect.ImmutableCollection;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMultimap;
 import org.gradle.api.InvalidUserDataException;
-import org.gradle.internal.execution.Result;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.WorkValidationException;
 import org.gradle.internal.reflect.MessageFormattingTypeValidationContext;
