+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/legacy/MarkSnapshottingInputsStartedStep.java
@@ -16,9 +16,9 @@
 
 package org.gradle.internal.execution.steps.legacy;
 
-import org.gradle.internal.execution.Result;
 import org.gradle.internal.execution.UnitOfWork;
 import org.gradle.internal.execution.steps.Context;
+import org.gradle.internal.execution.steps.Result;
 import org.gradle.internal.execution.steps.Step;
 
 /**
