+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/AssignWorkspaceStepTest.groovy
@@ -17,7 +17,6 @@
 package org.gradle.internal.execution.steps
 
 
-import org.gradle.internal.execution.Result
 import org.gradle.internal.execution.UnitOfWork
 import org.gradle.internal.execution.workspace.WorkspaceProvider
