+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/RecordOutputsStepTest.groovy
@@ -16,7 +16,7 @@
 
 package org.gradle.internal.execution.steps
 
-import org.gradle.internal.execution.CurrentSnapshotResult
+
 import org.gradle.internal.execution.history.OutputFilesRepository
 
 class RecordOutputsStepTest extends ContextInsensitiveStepSpec implements SnasphotterFixture {
