+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/ResolveInputChangesStepTest.groovy
@@ -17,7 +17,6 @@
 package org.gradle.internal.execution.steps
 
 
-import org.gradle.internal.execution.Result
 import org.gradle.internal.execution.UnitOfWork
 import org.gradle.internal.execution.history.changes.ExecutionStateChanges
 import org.gradle.internal.execution.history.changes.InputChangesInternal
